requirements.txt
Flask==2.3.3
pandas==2.0.3
openpyxl==3.1.2
google-generativeai==0.3.2